Add FileTestBase abstraction; Create CSVFileTest class with setUp method
