+++ appinventor/components/tests/com/google/appinventor/components/runtime/CSVFileTest.java
@@ -1,0 +1,15 @@
+package com.google.appinventor.components.runtime;
+
+import org.junit.Before;
+
+import static org.junit.Assert.*;
+
+public class CSVFileTest extends FileTestBase {
+  protected CSVFile csvFile;
+
+  @Before
+  public void setUp() {
+    super.setUp();
+    csvFile = new CSVFile(getForm());
+  }
+}
